Create VoiceChannelManager
Refactor all voice channel related code to a new dedicated class: VoiceChannelManager.

Remove voice state update event
This introduces the risk of race conditions where a user leaves and another joins.
Instead rely on the existing mechanism that will eventually clear the channel after it is empty for a while.