Fix deployment edit field order
Previosly the field order was based on the select menu. I.e. the fields would appear in the order you select the fields to edit.

Now the fields appear in a fixed order, the same order as new deployment modal.

Consolidate the 2 modals and remove the discrepencies around time placeholder and max length.