Update role emoji code
Fix bug in formatRoleEmoji where we always compare to FIRETEAM instead
of the input role.

Also rewrite some of the emoji code to be more error resistant. Due to changing offense to fireteam, we are getting errors trying to load an emoji that doesn't exist anymore.
Include a backwards compatible check for the old Offense role.

 #cleanup