Improve interaction logging
Log all interactions with more detail to make debugging easier.
When interactions start, log the user, the message (if one exists) and the interaction id.
When an interaction ends, log the interaction id for success.

Add logs at the end of each interaction to indicate what happened,
including details. E.g. when a deployment is created, log the deployment
name/id/message.