Proactively check for permissions
The discord permissions error does not specify what permissions are missing.
We proactively check for permissions and provide a meaningful and helpful message if permissiosn are missing before attempting an operation.

This does not protect us from permissions we need but not yet know about but we can add them when we figure out what perms we need.