Move interaction logging to after the inCachedGuild check
We don't currently support non cached interactions and i suspect that it might be causing errors around restarts of the bot since some of the properties are missing (e.g. `interaction.member`).

I don't yet fully understand the difference, but it seems that non cached interactions happen if the interaction started before the bot was restarted and completed after.