Add edit deployment logging
Make the link between the 2 interaction ids in edit deployment more explicit.
Move the selected values log ealier.
Add a log before editing the reply (this is one of the few places where we don't go through the utility functions for editing a reply and hence don't get the logs from there about this happening).

We have seen a recent bug where the show modal timed out but there doesn't seem to be a time gap between when the interaction begins and when we get the `Unknown interaction` error. I can't figure out what is going on from the logs. Adding some more logs in the hopes of having better visibility next time.