Check edit permissions in the transaction.
Currently the check is done before showing the modal, this can result in a race where the notice is sent after the modal is sent but before the edit is commited, sending the wrong information.

Performing the check inside the transaction means we can't edit a deployment if the notice is sent.

We keep the check before the modal is shown to prevent someone without permissions from getting the edit modal and wasting time trying to edit a deployment they don't have permissions to edit.

Issue #10