Replace awaitMessageComponent with a regular interaction flow
Instead of waiting for the message component we let the select menu interaction hook bring us back to the edit flow.
This removes the wait and might help with some of the unknown interaction errors we have been seeing.
Either way, this elimites the only place where we treat interactions differently than the others by waiting for something while processing a different interaction and ignoring the interaction callback for the select menu elsewhere.