Handle missing deployment messages
Trying to fetch a missing deployment message results in an exception and we don't really have anything to do with a deployment that has its message manually deleted, we treat it as deleted and ignore it in most places.

We also add wrappers for the operations since we might fetch the message but the message might get deleted before we attempt to edit or delete it.

We do seem to sometimes have a cached copy of a deleted message and because of that, we will successfully manage to fetch the message.
This is likely because we don't subscribe to the privilaged message content intent and hence our cache is out of date and contains deleted messages.
If this is something we want to handle, we either need to force an action (e.g. message.edit) or more likely, subscribe to the required intent to keep the cache up to date.